# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/manage:
    get:
      operationId: listManagedEvents
      tags: [events]
      summary: List events for account management
      responses:
        "200":
          description: All events with activation state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          description: Authentication required
        "403":
          description: Admin role required
  /events:
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a public event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid event
        "401":
          description: Authentication required
        "403":
          description: Admin role required
  /events/{eventId}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update a public event
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid event
        "401":
          description: Authentication required
        "403":
          description: Admin role required
        "404":
          description: Event not found
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete a public event
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "204":
          description: Deleted
        "401":
          description: Authentication required
        "403":
          description: Admin role required
        "404":
          description: Event not found
...
  /session/me:
    get:
      operationId: getCurrentUser
      tags: [session]
      summary: Get the current application user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "401":
          description: Authentication required
...
components:
  parameters:
...
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Event:
      type: object
      required: [id, title, message, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        message: { type: string }
        imageUrl: { type: ["string", "null"], format: uri }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EventInput:
      type: object
      required: [title, message]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        message: { type: string, minLength: 1, maxLength: 1000 }
        imageUrl: { type: ["string", "null"], format: uri }
        isActive: { type: boolean }
    EventUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        message: { type: string, minLength: 1, maxLength: 1000 }
        imageUrl: { type: ["string", "null"], format: uri }
        isActive: { type: boolean }
...
    AppUser:
      type: object
      required: [id, name, email, role, joinedAt, lastActiveAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        avatarUrl: { type: ["string", "null"], format: uri }
        role:
          type: string
          enum: [admin, moderator, member]
        joinedAt: { type: string, format: date-time }
        lastActiveAt: { type: string, format: date-time }
        profileVisibility:
          type: string
          enum: [private, public]
...